# Number refactorings with positive effects