I spent some time factoring number parsing, to get rid of the need to pass the writer to a function ... I haven't succeeded in that yet, but the small changes here seem to change some stuff so I thought I'd send it up. IMO it makes the flow a little clearer, which is the main reason to do it.

* Return error codes from parse_double(): this was actually the biggest jump. Now that we're just passing error codes all the way down to the caller, we get some savings!
* Make compute_float_64 return bool (and write to `double &d`): this simplifies our control flow usage a bit. This also gets rid of a "not quite bug" where one of the paths in compute_float_64 forgot to assign success = true, so you are forced to pre-assign it to true before calling.
* Collapses the code to remove leading zeroes into its own function (significant_digits()).

## Performance

This appears to make things better for files like canada.json, which has lots of numbers dispersed across many different arrays (it's an array of points), and worse for files like numbers.json, which has one massive array.  These numbers are vs. master, but note that this PR is relative to #1101 .

It also results in a general, seemingly significant, drop in branch misses on ARM for most files.

ARM gcc10.2 (ampere):

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| canada                           |  35172 | 811.2 | 781.4 |    3% | 996.4 | 987.7 |    0% |  48499 |  44052 |    9% |  91719 |  91769 |    0% | 5088797 | 5097982 |    0% |
| update-center                    |   8330 | 553.4 | 540.1 |    2% | 455.2 | 455.3 |    0% |  13610 |  13558 |    0% |  20787 |  20814 |    0% | 1008397 | 1021905 |   -1% |
| mesh.pretty                      |  24646 | 563.9 | 556.8 |    1% | 624.7 | 622.6 |    0% |  24094 |  23485 |    2% |  59090 |  59084 |    0% | 1971915 | 1974519 |    0% |
| random                           |   7976 | 708.8 | 702.1 |    0% | 636.1 | 636.1 |    0% |   7126 |   8713 |  -22% |  21598 |  21598 |    0% | 1235929 | 1242282 |    0% |
| apache_builds                    |   1988 | 474.3 | 470.8 |    0% | 411.5 | 411.5 |    0% |    730 |    717 |    1% |   4832 |   4822 |    0% | 201044 | 200786 |    0% |
| twitter                          |   9867 | 491.0 | 488.0 |    0% | 381.9 | 381.9 |    0% |  16676 |  16396 |    1% |  23394 |  23385 |    0% | 856794 | 857169 |    0% |
| citm_catalog                     |  26987 | 392.5 | 391.3 |    0% | 361.8 | 361.8 |    0% |   9374 |   8868 |    5% |  62694 |  62674 |    0% | 1818343 | 1816310 |    0% |
| gsoc-2018                        |  51997 | 359.6 | 358.9 |    0% | 237.6 | 237.6 |    0% |  38316 |  39331 |   -2% | 109503 | 109527 |    0% | 2739706 | 2752278 |    0% |
| twitterescaped                   |   8787 | 739.4 | 738.7 |    0% | 620.1 | 620.1 |    0% |  19863 |  19676 |    0% |  21319 |  21336 |    0% | 1562384 | 1560415 |    0% |
| github_events                    |   1017 | 461.0 | 461.1 |    0% | 357.0 | 357.0 |    0% |   1530 |   1505 |    1% |   2406 |   2401 |    0% |  87753 |  87173 |    0% |
| instruments                      |   3442 | 489.5 | 491.0 |    0% | 464.5 | 464.5 |    0% |   4318 |   4468 |   -3% |   8675 |   8670 |    0% | 340896 | 341213 |    0% |
| mesh                             |  11306 | 922.5 | 929.6 |    0% | 1047.7 | 1049.1 |    0% |  21104 |  18051 |   14% |  32339 |  32347 |    0% | 1648614 | 1649047 |    0% |
| marine_ik                        |  46616 | 861.4 | 868.9 |    0% | 968.3 | 969.4 |    0% |  95052 |  92301 |    2% | 134047 | 133978 |    0% | 7538697 | 7571395 |    0% |
| numbers                          |   2345 | 888.3 | 909.5 |   -2% | 873.9 | 878.2 |    0% |   3148 |   3239 |   -2% |   6021 |   6017 |    0% | 258282 | 258396 |    0% |

Haswell gcc9.2 (skylake):

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| canada                           |  35172 | 288.8 | 273.2 |    5% | 973.9 | 965.4 |    0% |  19401 |  18932 |    2% |  74999 |  69317 |    7% | 343822 | 345134 |    0% |
| marine_ik                        |  46616 | 294.5 | 281.3 |    4% | 919.6 | 921.7 |    0% |  52315 |  51872 |    0% | 224627 | 258374 |  -15% | 570302 | 575376 |    0% |
| mesh                             |  11306 | 308.9 | 297.5 |    3% | 1009.0 | 1010.8 |    0% |  10039 |  10142 |   -1% |      0 |      0 |    0% | 136713 | 136863 |    0% |
| mesh.pretty                      |  24646 | 179.5 | 175.2 |    2% | 592.0 | 589.8 |    0% |  13921 |  13507 |    2% |   1496 |   1783 |  -19% | 194445 | 194803 |    0% |
| update-center                    |   8330 | 112.0 | 109.4 |    2% | 335.5 | 335.5 |    0% |   5702 |   5549 |    2% |      0 |      0 |    0% |  84324 |  84109 |    0% |
| numbers                          |   2345 | 266.9 | 261.2 |    2% | 819.3 | 823.6 |    0% |   2138 |   2204 |   -3% |      0 |      0 |    0% |  18414 |  18258 |    0% |
| instruments                      |   3442 | 105.2 | 103.3 |    1% | 378.5 | 380.0 |    0% |    459 |    536 |  -16% |      0 |      0 |    0% |  30156 |  30633 |   -1% |
| random                           |   7976 | 141.6 | 139.1 |    1% | 494.6 | 495.2 |    0% |   2040 |   1970 |    3% |      0 |      0 |    0% |  91854 |  91713 |    0% |
| twitter                          |   9867 |  91.7 |  90.3 |    1% | 290.2 | 290.5 |    0% |   3712 |   3622 |    2% |      0 |      0 |    0% |  83239 |  83090 |    0% |
| gsoc-2018                        |  51997 |  71.8 |  70.9 |    1% | 167.0 | 167.0 |    0% |  30701 |  31347 |   -2% |  86086 | 129569 |  -50% | 362910 | 366977 |   -1% |
| apache_builds                    |   1988 |  85.9 |  84.9 |    1% | 304.7 | 304.7 |    0% |    108 |    106 |    1% |      0 |      0 |    0% |  13691 |  14007 |   -2% |
| github_events                    |   1017 |  79.5 |  78.7 |    1% | 264.5 | 264.6 |    0% |    115 |    137 |  -19% |      0 |      0 |    0% |   4565 |   3515 |   23% |
| twitterescaped                   |   8787 | 184.5 | 184.3 |    0% | 489.9 | 490.2 |    0% |   5017 |   5197 |   -3% |      0 |      0 |    0% |  79442 |  79276 |    0% |
| citm_catalog                     |  26987 |  80.2 |  80.4 |    0% | 296.3 | 296.8 |    0% |   1735 |   1730 |    0% |   2830 |   4712 |  -66% | 193900 | 194843 |    0% |